Add files via upload
The dataset contains a subset of the Enron email corpus which has been categorized by researchers. This dataset contains a column, 'company_business,' which is a binary flag indicating whether the email pertains to company business or not. 

I am using this dataset to design a model to assist with document review. The purpose of this model is to find all the documents that are related to company business. Documents that the model labels as relevant will be stored, and the rest will be discarded. The company estimates that a document that should have been stored but was discarded costs them $200 in lost potential value, and a document that should have been discarded but was stored costs $40 in administration, storage, and review.

Building and optimizing a model using a performance criteria that takes into account these costs. Demonstrating an attempt use multiple machine learning techniques. Once I found the optimal model, I used its performance to provide an estimated error cost per 100 documents.